增加编辑接口

+++ linkwe-wecom/src/main/java/com/linkwechat/wecom/service/IWeTaskFissionService.java
@@ -43,7 +43,7 @@ public interface IWeTaskFissionService {
      * @param weTaskFission 任务宝
      * @return 结果
      */
-    public int updateWeTaskFission(WeTaskFission weTaskFission);
+    public Long updateWeTaskFission(WeTaskFission weTaskFission);
 
     /**
      * 批量删除任务宝
